# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "This is my website cool API",
        "description": "Api description...",
        "termsOfService": "",
        "contact": {
            "email": "contact@mysite.com"
        },
        "license": {
            "name": "Private License",
            "url": "URL to the license"
        },
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/contact": {
            "post": {
                "summary": "Contact form",
                "description": "Contact form for users",
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "User name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "The email for contact",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "subject",
                        "in": "formData",
                        "description": "Subject of mail",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "message",
                        "in": "formData",
                        "description": "Message which user sends",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "game/all": {
            "get": {
                "summary": "Get list of all games",
                "description": "Returns list of all games",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "mod/all/{game_id}": {
            "get": {
                "summary": "Get list of all mods",
                "description": "Returns list of all mods",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/order": {
            "post": {
                "summary": "Make order",
                "description": "Make order",
                "parameters": [
                    {
                        "name": "gameId",
                        "in": "formData",
                        "description": "Which game user wants",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "machineId",
                        "in": "formData",
                        "description": "Where to be stored",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "slots",
                        "in": "formData",
                        "description": "How many slots",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "months",
                        "in": "formData",
                        "description": "How long to be active",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "serverName",
                        "in": "formData",
                        "description": "Server Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "modId",
                        "in": "formData",
                        "description": "Which mod",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "server/all": {
            "get": {
                "summary": "Get list of all servers",
                "description": "Returns list of all servers",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/login": {
            "post": {
                "summary": "Login with jwt",
                "description": "Returns jwt and all user data",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "The email for login",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "The password for login",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/register": {
            "post": {
                "summary": "Registration form",
                "description": "Registration form with cart",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "formData",
                        "description": "First name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "last_name",
                        "in": "formData",
                        "description": "Last name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "username",
                        "in": "formData",
                        "description": "Username",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Neznam123!@",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password_confirmation",
                        "in": "formData",
                        "description": "Neznam123!@",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "pin_code",
                        "in": "formData",
                        "description": "Pin code",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "country",
                        "in": "formData",
                        "description": "Country",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Api Not found!"
                    }
                }
            }
        },
        "/user": {
            "get": {
                "summary": "Get authenticated user",
                "description": "Returns authenticated user",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/user/profile/update": {
            "post": {
                "summary": "Update user profile",
                "description": "Update user profile",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "formData",
                        "description": "First name",
                        "type": "string"
                    },
                    {
                        "name": "last_name",
                        "in": "formData",
                        "description": "Last name",
                        "type": "string"
                    },
                    {
                        "name": "username",
                        "in": "formData",
                        "description": "Username",
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email",
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Neznam123!@",
                        "type": "string"
                    },
                    {
                        "name": "pin_code",
                        "in": "formData",
                        "description": "Pin code",
                        "type": "string"
                    },
                    {
                        "name": "country",
                        "in": "formData",
                        "description": "Country",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Api Not found!"
                    }
                }
            }
        }
    },
    "definitions": {},
    "externalDocs": {
        "description": "Find out more about my website",
        "url": "http..."
    }
}